refactor: ChannelCategory to Channel
